Revert "Move SessionRepository classes to repository package"
