+++ src/main/java/org/cbioportal/session_service/domain/SessionRepository.java
@@ -32,9 +32,10 @@
 
 package org.cbioportal.session_service.domain;
 
-import org.cbioportal.session_service.repository.SessionRepositoryCustom;
+import org.cbioportal.session_service.domain.internal.SessionRepositoryCustom;
  
 import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.data.mongodb.repository.Query;
 
 /**
  * @author Manda Wilson
