+++ src/main/java/org/cbioportal/session_service/domain/internal/SessionRepositoryCustom.java
@@ -30,7 +30,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package org.cbioportal.session_service.repository;
+package org.cbioportal.session_service.domain.internal;
 
 import org.cbioportal.session_service.domain.Session;
 import org.cbioportal.session_service.domain.SessionType;
